Move splitext_improved to module
