+++ ell1fit/ell1fit.py
@@ -7,6 +7,7 @@ import logging
 
 import matplotlib.pyplot as plt
 import numpy as np
+from ell1fit import splitext_improved
 from hendrics.io import load_events
 from stingray.pulse.pulsar import z_n_events, z_n_binned_events
 from stingray.stats import a_from_ssig
@@ -188,39 +189,6 @@ def pf_weight_versus_energy(times, energies, parameters, nbin=32, nharm=1, toler
         weights.append(weight_func(local_energies))
 
     return weights
-
-
-def splitext_improved(path):
-    """
-    Examples
-    --------
-    >>> np.all(splitext_improved("a.tar.gz") ==  ('a', '.tar.gz'))
-    True
-    >>> np.all(splitext_improved("a.tar") ==  ('a', '.tar'))
-    True
-    >>> path_with_dirs = os.path.join("a.f", "a.tar")
-    >>> path_without_ext = os.path.join("a.f", "a")
-    >>> np.all(splitext_improved(path_with_dirs) ==  (path_without_ext, '.tar'))
-    True
-    >>> path_with_dirs = os.path.join("a.a.a.f", "a.tar.gz")
-    >>> path_without_ext = os.path.join("a.a.a.f", "a")
-    >>> np.all(splitext_improved(path_with_dirs) ==  (path_without_ext, '.tar.gz'))
-    True
-    >>> path_with_dirs = os.path.join("a.a.a.f", "a.1.tar")
-    >>> path_without_ext = os.path.join("a.a.a.f", "a.1")
-    >>> np.all(splitext_improved(path_with_dirs) ==  (path_without_ext, '.tar'))
-    True
-    """
-    import os
-
-    dir, file = os.path.split(path)
-
-    if len(file.split(".")) > 2 and file.endswith(".gz"):
-        froot, ext = file.split(".")[0], "." + ".".join(file.split(".")[-2:])
-    else:
-        froot, ext = os.path.splitext(file)
-
-    return os.path.join(dir, froot), ext
 
 
 @njit
